lite/requirements.txt
# Prime Lite Requirements
# Minimal dependencies for personal AI agent mode

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
python-keyring>=24.3.0

# Config
pyyaml>=6.0.1
python-dotenv>=1.0.0

# HTTP
httpx>=0.26.0
aiohttp>=3.9.0

# Telegram
python-telegram-bot>=20.8

# Utils
pydantic>=2.5.0
pydantic-settings>=2.1.0
